% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{collect_funds_small_large}
\alias{collect_funds_small_large}
\title{FUNCTION: collect_funds_small_large}
\usage{
collect_funds_small_large(d)
}
\description{
Group data by initial funds, the small and large date ranges for calculating simple moving averages.
}

