% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{display_algorithm_worth}
\alias{display_algorithm_worth}
\title{FUNCTION: display_algorithm_worth}
\usage{
display_algorithm_worth(d)
}
\description{
A function for displaying the worth of each algorithm for many configuration scenarios.
}

