% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_algorithm_data}
\alias{read_algorithm_data}
\title{FUNCTION: read_algorithm_data}
\usage{
read_algorithm_data(d)
}
\description{
Read the data collected from buying and selling using greedy and random algorithms.
}

