% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{summarize_algorithm_worth}
\alias{summarize_algorithm_worth}
\title{FUNCTION: summarize_algorithm_worth}
\usage{
summarize_algorithm_worth(d)
}
\description{
A function for determining which algorithm is better on average at generating more 'worth' in one month's time.
}

